move realization to cpp files
